// web/api-gen/models/MetricsStoreSpec.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MetricsStorePrometheus } from './MetricsStorePrometheus';
import type { MetricsStoreVictoriaMetrics } from './MetricsStoreVictoriaMetrics';
export type MetricsStoreSpec = {
    prometheus?: MetricsStorePrometheus;
    victoriametrics?: MetricsStoreVictoriaMetrics;
};